pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Raghava B <raghava.b@integraconnect.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
openai = "^1.19.0"
azure-identity = "^1.16.0"
sqlmodel = "^0.0.16"
pyodbc = "^5.1.0"
pydantic = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
